Проект в рамках учебы. Создание и описание примитивного бота, способного уведомлять о каких-либо напоминаниях.

+++ src/main/resources/liquibase/scripts/botScripts.sql
@@ -3,7 +3,7 @@
 -- changeset rvoroshnin:1
 CREATE TABLE notification_task (
     id SERIAL PRIMARY KEY,
-    chat_id SERIAL,
+    chat_id BIGINT,
     message_text TEXT,
     notification_date TIMESTAMP WITH TIME ZONE
 )
